+++ Deeploy/Targets/PULPOpen/Templates/FloatGemmTemplate.py
@@ -24,9 +24,18 @@ for(uint32_t i=0; i<${batch}; i++){
         ${transB}
     );
 
+    % if A_batched:
     ref_${data_out}_${A} += ${M} * ${N};
+    % endif
+
+    % if B_batched:
     ref_${data_out}_${B} += ${N} * ${O};
+    % endif
+
+    % if C_batched:
     ref_${data_out}_${C} += ${M} * ${O};
+    % endif
+
     ref_${data_out}_${data_out} += ${M} * ${O};
 }
 """)