+++ Deeploy/Targets/PULPOpen/Templates/ReshapeTemplate.py
@@ -25,10 +25,11 @@
 
 from typing import Dict, List, Tuple
 
-from Deeploy.DeeployTypes import NetworkContext, NodeTemplate, OperatorRepresentation
+from Deeploy.DeeployTypes import NetworkContext, OperatorRepresentation, VariableBuffer
+from Deeploy.Targets.Generic.Templates.ReshapeTemplate import _ReshapeTemplate as _GenericReshapeTemplate
 
 
-class _ReshapeTemplate(NodeTemplate):
+class _ReshapeTemplate(_GenericReshapeTemplate):
 
     def __init__(self, templateStr):
         super().__init__(templateStr)
@@ -36,19 +37,18 @@ class _ReshapeTemplate(NodeTemplate):
     def alignToContext(self, ctxt: NetworkContext,
                        operatorRepresentation: OperatorRepresentation) -> Tuple[NetworkContext, Dict, List[str]]:
 
-        # SCHEREMO: Selectively mark 'indices' dead, since we don't need them
-        if 'indices' in operatorRepresentation.keys():
-            ctxt.globalObjects[operatorRepresentation['indices']]._deploy = False
-            ctxt.globalObjects[operatorRepresentation['indices']]._live = False
+        ctxt, operatorRepresentation, _ = super().alignToContext(ctxt, operatorRepresentation)
 
-        # Same for "shape"
-        if "shape" in operatorRepresentation.keys():
-            ctxt.globalObjects[operatorRepresentation["shape"]]._deploy = False
-            ctxt.globalObjects[operatorRepresentation["shape"]]._live = False
+        # Get buffers
+        bufferIn = ctxt.lookup(operatorRepresentation['data_in'])
+        assert isinstance(bufferIn, VariableBuffer)
 
-        inBuffer = ctxt.lookup(operatorRepresentation['data_in'])
-        outBuffer = ctxt.lookup(operatorRepresentation['data_out'])
-        outBuffer._alias = inBuffer.name
+        bufferOut = ctxt.lookup(operatorRepresentation['data_out'])
+        assert isinstance(bufferOut, VariableBuffer)
+
+        # HACK: Tiling wasn't updated in the Fix aliasing PR so we have to still
+        #       set the _alias argument
+        bufferOut._alias = bufferIn.name
 
         return ctxt, operatorRepresentation, []