+++ DeeployTest/testUtils/testRunner.py
@@ -342,6 +342,10 @@ class TestRunner():
             generation_script = "generateNetwork.py"
 
         command = f"python {generation_script} -d {self._dir_gen} -t {self._dir_test} -p {self._platform} {self.gen_args}"
+
+        if self._platform in ["Siracusa", "Siracusa_w_neureka"]:
+            command += f" --cores={self._args.cores}"
+
         command += self._argument_parser.generate_cmd_args()
 
         log.debug(f"[TestRunner] Generation Command: {command}")